+++ CHANGELOG.md
@@ -1,21 +1,34 @@
 # Change Logs
 
+## [v0.2.4] - 2024-02-08
+
+### Added
+
+- Get the Organization's README.md
+
+### Changed
+
+- Update domain
+- Update side drawer UX
+- Update article style
+- Other fixed
+
 ## [v0.2.3] - 2021-11-10
 
-## Changed
+### Changed
 
 - purgeCSS disabled
 
 
 ## [v0.2.2] - 2021-11-10
 
-## Changed
+### Changed
 
 - Changed from npm to yarn
 - Import icons and components directly
 - Improved performance
 
-## Fixed
+### Fixed
 
 - Other Fixed